+++ src/app/clusters/window-covering-server/app_config_dependent_sources.cmake
@@ -16,6 +16,8 @@
 TARGET_SOURCES(
   ${APP_TARGET}
   PRIVATE
+    "${CLUSTER_DIR}/WindowCoveringCluster.h"
+    "${CLUSTER_DIR}/WindowCoveringDelegate.h"
     "${CLUSTER_DIR}/window-covering-delegate.h"
     "${CLUSTER_DIR}/WindowCoveringCluster.cpp"
     "${CLUSTER_DIR}/window-covering-server.h"
